Move QsbSearchProvider down into `providers` package
