+++ lawnchair/src/app/lawnchair/qsb/QsbLayout.kt
@@ -15,6 +15,7 @@ import androidx.core.view.isVisible
 import app.lawnchair.launcher
 import app.lawnchair.preferences.PreferenceManager
 import app.lawnchair.preferences2.PreferenceManager2
+import app.lawnchair.qsb.providers.QsbSearchProvider
 import app.lawnchair.util.viewAttachedScope
 import com.android.launcher3.BaseActivity
 import com.android.launcher3.DeviceProfile
